Renaming file and its occurrences in the project to make it more pythonic
